requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
coverage==7.6.4
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Testing==0.8.1
gunicorn==23.0.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
pluggy==1.5.0
pytest==8.3.3
python-dotenv==1.0.1
requests==2.32.3
tomli==2.0.2
urllib3==2.2.3
Werkzeug==3.0.4
